requirements.txt
notion_client==2.2.1
pandas==1.5.3
python-dotenv==1.0.1